Check if access token is present

+++ packages/bygger-backend/src/routers/api/recipients/recipients.ts
@@ -21,7 +21,10 @@ const get: RequestHandler = async (req, res, next) => {
 };
 
 const post: RequestHandler = async (req, res, next) => {
-  const accessToken = req.get('AzureAccessToken');
+  const accessToken = req.headers.AzureAccessToken as string;
+  if (!accessToken) {
+    next(new Error('No access token provided'));
+  }
   try {
     const recipient = await recipientService.post(req.body, accessToken);
     res.json(recipient);
@@ -32,7 +35,10 @@ const post: RequestHandler = async (req, res, next) => {
 
 const put: RequestHandler = async (req, res, next) => {
   const { recipientId } = req.params;
-  const accessToken = req.get('AzureAccessToken');
+  const accessToken = req.headers.AzureAccessToken as string;
+  if (!accessToken) {
+    next(new Error('No access token provided'));
+  }
   try {
     const recipient = await recipientService.put(recipientId, req.body, accessToken);
     res.json(recipient);
@@ -43,7 +49,10 @@ const put: RequestHandler = async (req, res, next) => {
 
 const deleteRecipient: RequestHandler = async (req, res, next) => {
   const { recipientId } = req.params;
-  const accessToken = req.get('AzureAccessToken');
+  const accessToken = req.headers.AzureAccessToken as string;
+  if (!accessToken) {
+    next(new Error('No access token provided'));
+  }
   try {
     await recipientService.delete(recipientId, accessToken);
     res.end();
